소모임 폼 수정 , 댓글 폼 수정, 파트너별점 추가

+++ ForPets/src/main/java/com/forpets/biz/partnerReview/Impl/PartnerReviewDAO.java
@@ -87,8 +87,8 @@ public class PartnerReviewDAO {
 		return jdbcTemplate.query(PT_REVIEW, previewRowMapper, part_id);
 	}
 	
-	public Double avg(String part_id) {
-		Double avg = jdbcTemplate.queryForObject(AVG, double.class, part_id);
+	public int avg(String part_id) {
+		int avg = jdbcTemplate.queryForObject(AVG, Integer.class, part_id);
 		return avg;
 	}
